added create, update and remove post functions
