+++ index.html
@@ -8,7 +8,7 @@
 	<title>PlantPals | Login</title>
 	<meta name="description" content="This is the homepage for PlantPals where you can sign up or login">
 	<link rel="stylesheet" href="./dist/css/index.css">
-	<script type="module" src="index.js"></script>
+	<script defer type="module" src="./js/index.mjs"></script>
 	<script src="https://kit.fontawesome.com/727f4fec37.js" crossorigin="anonymous"></script>
 </head>
