--- index.js
@@ -1,3 +1,0 @@
-import { homePage } from "./js/routes/home.js";
-
-homePage()
