+++ js/api/headers.mjs
@@ -1,5 +1,5 @@
-import { load } from "../storage/load.js";
-import { API_KEY } from "./constants.js";
+import { load } from "../storage/load.mjs";
+import { API_KEY } from "./constants.mjs";
 
 export function headers(hasBody = false) {
 const headers = new Headers();
