+++ js/api/posts/index.mjs
@@ -1,0 +1,4 @@
+export * from "./delete.mjs";
+export * from "./read.mjs";
+export * from "./update.mjs";
+export * from "./create.mjs";
