+++ js/index.mjs
@@ -1,0 +1,6 @@
+import { homePage } from "./routes/home.mjs";
+
+homePage()
+
+
+
