--- js/routes/home.js
@@ -1,6 +1,0 @@
-import { setAuthListener } from "../ui/listeners/auth.js";
-
-export async function homePage() {
-	setAuthListener()
-}
-
