+++ js/routes/home.mjs
@@ -1,0 +1,6 @@
+import { setAuthListener } from "../ui/listeners/auth.mjs";
+
+export async function homePage() {
+	setAuthListener()
+}
+
