+++ js/ui/listeners/auth.mjs
@@ -1,5 +1,6 @@
-import { onAuth } from "../events/onAuth.js";
+import { onAuth } from "../events/onAuth.mjs";
 
 export function setAuthListener() {
 	document.forms.auth.addEventListener("submit", onAuth);
 }
+
