+++ post/index.html
@@ -1,0 +1,174 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+	<meta charset="UTF-8">
+	<meta http-equiv="X-UA-Compatible" content="IE=edge">
+	<meta name="viewport" content="width=device-width, initial-scale=1.0">
+	<title>PlantPals | Post</title>
+	<meta name="description"
+		content="Here you can see the posts that other peolpe you follow have postet in PlantPals, and you can also write a new post to your own profile">
+	<link rel="stylesheet" href="../dist/css/index.css">
+	<!-- <script defer src="node_modules/bootstrap/dist/js/bootstrap.js"></script> -->
+	<script src="https://kit.fontawesome.com/727f4fec37.js" crossorigin="anonymous"></script>
+</head>
+
+<body class="min-vh-100">
+	<header class="bg-primary">
+		<nav class="navbar navbar-light justify-content-evenly">
+			<div>
+				<h1>PlantPals</h1>
+			</div>
+			<form class="form-inline my-2 my-lg-0">
+				<input class="form-control mr-sm-2" type="search" placeholder="Search" aria-label="Search">
+				<button class="btn btn-outline-success my-2 my-sm-0" type="submit">Search</button>
+			</form>
+			<p>Log out</p>
+		</nav>
+	</header>
+	<!-- sidebar start -->
+	<div class="container-fluid">
+		<div class="row flex-nowrap">
+			<div class="col-auto col-md-3 col-xl-2 px-sm-2 px-0 bg-white">
+				<div class="d-flex flex-column align-items-center align-items-sm-start px-3 pt-2 text-white min-vh-100">
+
+					<ul class="nav nav-pills flex-column mb-sm-auto mb-0 align-items-center align-items-sm-start"
+						id="menu">
+						<li class="nav-item">
+							<a href="../posts/index.html" class="nav-link align-middle px-0">
+								<i class="fs-4 bi-house"></i> <span class="ms-1 d-none d-sm-inline">Home</span>
+							</a>
+						</li>
+
+						<li class="nav-item">
+							<a href="../posts/index.html" class="nav-link align-middle px-0">
+								<i class="fs-4 bi-house"></i> <span class="ms-1 d-none d-sm-inline">Add new post</span>
+							</a>
+						</li>
+						<li class="nav-item">
+							<a href="#" class="nav-link align-middle px-0">
+								<i class="fs-4 bi-house"></i> <span class="ms-1 d-none d-sm-inline">Settings</span>
+							</a>
+						</li>
+						<li class="nav-item">
+							<a href="#" class="nav-link align-middle px-0">
+								<i class="fs-4 bi-house"></i> <span class="ms-1 d-none d-sm-inline">Notifications</span>
+							</a>
+						</li>
+						<li class="nav-item">
+							<a href="../profile/index.html" class="nav-link align-middle px-0">
+								<i class="fs-4 bi-house"></i> <span class="ms-1 d-none d-sm-inline">Profile</span>
+							</a>
+						</li>
+					</ul>
+				</div>
+			</div>
+			<div class="col py-3">
+
+				<!-- sidebar end -->
+				<main>
+					<!-- New post form start-->
+					<div class="container">
+						<div class="row col-md-8 col-lg-4 col-xl-5">
+							<h2 class="my-5">For you</h2>
+							<p class="m-0">+ Add new post</p>
+							<form class="was-validated">
+								<div class="mb-3">
+									<label for="validationTextarea" class="form-label"></label>
+									<textarea class="form-control" id="validationTextarea"
+										placeholder="Required example textarea" required></textarea>
+									<div class="invalid-feedback">
+										Add text
+									</div>
+								</div>
+
+								<div class="mb-3">
+									<input type="file" class="form-control" aria-label="file example" required>
+									<div class="invalid-feedback">Please select a file</div>
+								</div>
+
+								<div class="mb-3">
+									<button class="btn btn-primary" type="submit" disabled>Post</button>
+								</div>
+							</form>
+						</div>
+					</div>
+					<!-- New post form end -->
+
+					<!-- card 1 -->
+					<div id="feed" class="row justify-content-center">
+						<div class="card shadow col-10 col-lg-3 mx-3 mb-3 d-flex">
+							<div class="flex-fill">
+								<div class="card-body">
+									<h5 class="card-title">plant_vibes</h5>
+								</div>
+								<div class="card-footer text-muted">
+									2 days ago
+								</div>
+								<img src="../images/small_plant.jpg" alt="small growing plant" width="100%">
+								<div class="card-body">
+									<p class="card-text">After five weeks of patient nurturing, the once dormant seed
+										has
+										undergone
+										a
+										remarkable
+										transformation. From a tiny, unassuming kernel, it has sprouted into a vibrant
+										and
+										delicate
+										sapling, reaching towards the sunlight with newfound vitality.</p>
+									<p>
+										@plant_vibes #plants #growing #planting #ecoliving
+									</p>
+								</div>
+							</div>
+							<div class="row">
+								<p>Like</p>
+								<p>Comment</p>
+								<p>Share</p>
+							</div>
+						</div>
+						<!-- card 2 -->
+
+						<!-- card 10  -->
+						<div class="card shadow col-10 col-lg-3 mx-3 mb-3 d-flex">
+							<div class="flex-fill">
+								<div class="card-body">
+									<h5 class="card-title">green_garden</h5>
+								</div>
+								<div class="card-footer text-muted">
+									3 days ago
+								</div>
+								<img src="../images/garden.jpg" alt="a beautyful garden" width="100%">
+								<div class="card-body">
+									<p class="card-text">
+										In the heart of every garden, whispers of tranquility unfold as each petal
+										unfurls,
+										telling
+										stories of growth, resilience, and the timeless beauty of the natural world.
+									</p>
+									<p>
+										#GardenStories #BlossomingBeauty #ZenRetreat
+									</p>
+								</div>
+								<div>
+									<p>Like</p>
+									<p>Comment</p>
+									<p>Share</p>
+								</div>
+							</div>
+						</div>
+					</div>
+			</div>
+			<!-- feed posts end -->
+			</main>
+
+		</div>
+	</div>
+	</div>
+	<footer class="bg-primary p-5">
+
+	</footer>
+
+</body>
+
+</html>
