+++ posts/index.html
@@ -35,17 +35,13 @@
 					<ul class="nav nav-pills flex-column mb-sm-auto mb-0 align-items-center align-items-sm-start"
 						id="menu">
 						<li class="nav-item">
-							<a href="#" class="nav-link align-middle px-0">
-								<i class="fs-4 bi-house"></i> <span class="ms-1 d-none d-sm-inline">Home</span>
+							<a href="../profile/index.html" class="nav-link align-middle px-0">
+								<i class="fs-4 bi-house"></i> <span class="ms-1 d-none d-sm-inline">Profile</span>
 							</a>
 						</li>
+
 						<li class="nav-item">
-							<a href="#" class="nav-link align-middle px-0">
-								<i class="fs-4 bi-house"></i> <span class="ms-1 d-none d-sm-inline">Search</span>
-							</a>
-						</li>
-						<li class="nav-item">
-							<a href="#" class="nav-link align-middle px-0">
+							<a href="../posts/index.html" class="nav-link align-middle px-0">
 								<i class="fs-4 bi-house"></i> <span class="ms-1 d-none d-sm-inline">Add new post</span>
 							</a>
 						</li>
@@ -57,11 +53,6 @@
 						<li class="nav-item">
 							<a href="#" class="nav-link align-middle px-0">
 								<i class="fs-4 bi-house"></i> <span class="ms-1 d-none d-sm-inline">Notifications</span>
-							</a>
-						</li>
-						<li class="nav-item">
-							<a href="#" class="nav-link align-middle px-0">
-								<i class="fs-4 bi-house"></i> <span class="ms-1 d-none d-sm-inline">Profile</span>
 							</a>
 						</li>
 					</ul>
