+++ register/index.html
@@ -8,7 +8,7 @@
 	<title>PlantPals | Sign up</title>
 	<meta name="description" content="This is the page where you can register and sign up for an account">
 	<link rel="stylesheet" href="../dist/css/index.css">
-	<script type="module" src="../index.js"></script>
+	<script defer type="module" src="../js/index.mjs"></script>
 	<script src="https://kit.fontawesome.com/727f4fec37.js" crossorigin="anonymous"></script>
 
 <body class="bg-primary min-vh-100">
